+++ src/test/php/peer/stomp/unittest/FrameFromWireTest.class.php
@@ -2,6 +2,7 @@
   
 use io\streams\MemoryInputStream;
 use io\streams\StringReader;
+use peer\stomp\Header;
 use peer\stomp\frame\MessageFrame;
 use unittest\TestCase;
 
@@ -21,13 +22,25 @@ class FrameFromWireTest extends TestCase {
 
   #[@test]
   public function frame_with_content_length() {
-    $frame= $this->frame("message-id: 1\ncontent-length: 4\n\nbody\0");
+    $frame= $this->frame("message-id:1\ncontent-length:4\n\nbody\0");
     $this->assertEquals('body', $frame->getBody());
   }
 
   #[@test]
   public function frame_without_content_length() {
-    $frame= $this->frame("message-id: 1\n\nbody\0");
+    $frame= $this->frame("message-id:1\n\nbody\0");
     $this->assertEquals('body', $frame->getBody());
+  }
+
+  #[@test]
+  public function frame_with_zero_content_length() {
+    $frame= $this->frame("message-id:1\ncontent-length:0\n\n\0");
+    $this->assertEquals('', $frame->getBody());
+  }
+
+  #[@test]
+  public function message_id_header() {
+    $frame= $this->frame("message-id:1\ncontent-length:4\n\nbody\0");
+    $this->assertEquals('1', $frame->getHeader(Header::MESSAGEID));
   }
 }