+++ ChangeLog.md
@@ -3,6 +3,8 @@ stomp ChangeLog
 
 ## ?.?.? / ????-??-??
 
+* Fixed reading frames with `content-length:0` - @thekid
+
 ## 10.0.0 / 2018-08-24
 
 * Made compatible with `xp-framework/logging` version 9.0.0 - @thekid